+++ Blog.Core.Extensions/ServiceExtensions/Authentication_Ids4Setup.cs
@@ -1,0 +1,40 @@
+﻿using Blog.Core.AuthHelper;
+using Blog.Core.Common;
+using Microsoft.AspNetCore.Authentication;
+using Microsoft.AspNetCore.Authentication.JwtBearer;
+using Microsoft.AspNetCore.Builder;
+using Microsoft.Extensions.DependencyInjection;
+using System;
+
+namespace Blog.Core.Extensions
+{
+    /// <summary>
+    /// Ids4权限 认证服务
+    /// </summary>
+    public static class Authentication_Ids4Setup
+    {
+        public static void AddAuthentication_Ids4Setup(this IServiceCollection services)
+        {
+            if (services == null) throw new ArgumentNullException(nameof(services));
+
+
+            // 添加Identityserver4认证
+            services.AddAuthentication(o =>
+            {
+                o.DefaultScheme = JwtBearerDefaults.AuthenticationScheme;
+                o.DefaultChallengeScheme = nameof(ApiResponseHandler);
+                o.DefaultForbidScheme = nameof(ApiResponseHandler);
+            })
+            .AddIdentityServerAuthentication(options =>
+            {
+                options.Authority = Appsettings.app(new string[] { "Startup", "IdentityServer4", "AuthorizationUrl" });
+                options.RequireHttpsMetadata = false;
+                options.ApiName = Appsettings.app(new string[] { "Startup", "IdentityServer4", "ApiName" });
+                options.SupportedTokens = IdentityServer4.AccessTokenValidation.SupportedTokens.Jwt;
+                options.ApiSecret = "api_secret";
+
+            })
+            .AddScheme<AuthenticationSchemeOptions, ApiResponseHandler>(nameof(ApiResponseHandler), o => { });
+        }
+    }
+}
